Move InstructionsPanel into its own file
