+++ apps/src/lab2/views/components/Instructions.tsx
@@ -1,27 +1,17 @@
-import {Button} from '@code-dot-org/component-library/button';
 import {useTheme} from '@code-dot-org/component-library/common/contexts';
-import classNames from 'classnames';
-import React, {useEffect, useRef} from 'react';
+import React from 'react';
 import {useSelector} from 'react-redux';
 
-import InstructorsOnly from '@cdo/apps/code-studio/components/InstructorsOnly';
 import {nextLevelId} from '@cdo/apps/code-studio/progressReduxSelectors';
 import {queryParams} from '@cdo/apps/code-studio/utils';
-import {LevelPredictSettings} from '@cdo/apps/lab2/levelEditors/types';
 import continueOrFinishLesson from '@cdo/apps/lab2/progress/continueOrFinishLesson';
 import {
   isPredictAnswerLocked,
   setPredictResponse,
 } from '@cdo/apps/lab2/redux/predictLevelRedux';
-import EnhancedSafeMarkdown from '@cdo/apps/templates/EnhancedSafeMarkdown';
-import {commonI18n} from '@cdo/apps/types/locale';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
 
-import PredictQuestion from './PredictQuestion';
-import PredictSummary from './PredictSummary';
-import TextToSpeech from './TextToSpeech';
-
-import moduleStyles from './instructions.module.scss';
+import InstructionsPanel from './InstructionsPanel';
 
 interface InstructionsProps {
   /** If the instructions panel should be rendered vertically or horizontally. Defaults to vertical. */
@@ -112,231 +102,4 @@ const Instructions: React.FunctionComponent<InstructionsProps> = ({
     />
   );
 };
-
-interface InstructionsPanelProps {
-  /** Primary instructions text to display. */
-  text: string;
-  /** Optional message to display under the main text. This is typically a validation message. */
-  message?: string;
-  /** Key for rendering the optional message. A unique value ensures the appearance animation shows. */
-  messageIndex?: number;
-  /** If the instructions panel should be rendered vertically or horizontally. Defaults to vertical. */
-  layout?: 'vertical' | 'horizontal';
-  /** Display theme. Defaults to Dark. */
-  theme?: 'Dark' | 'Light';
-  /**
-   * A callback when the user clicks on clickable text.
-   */
-  handleInstructionsTextClick?: (id: string) => void;
-  predictSettings?: LevelPredictSettings;
-  predictResponse?: string;
-  setPredictResponse: (response: string) => void;
-  predictAnswerLocked: boolean;
-  /** Optional classname for the container */
-  className?: string;
-  offerBrowserTts?: boolean;
-  canShowNextButton: boolean;
-  hasNextLevel: boolean;
-  useSecondaryFinishButton: boolean;
-  onContinueOrFinish: () => void;
-  bottomComponent?: React.ReactNode;
-  noAnimation?: boolean;
-}
-
-/**
- * Renders the instructions panel view. This was initially set up as a separate component
- * so that it could be used without the Lab2 redux integration if necessary.
- * If the level is a predict level, the predict reset button now uses redux, as it needs
- * multiple unique redux values and there isn't a clear use case for having no redux integration
- * anymore.
- * TODO: Determine if we need this separate component anymore, or if we can merge this into Instructions.
- * https://codedotorg.atlassian.net/browse/CT-671
- */
-export const InstructionsPanel: React.FunctionComponent<
-  InstructionsPanelProps
-> = ({
-  text,
-  message,
-  messageIndex,
-  layout = 'vertical',
-  theme = 'Dark',
-  handleInstructionsTextClick,
-  predictSettings,
-  predictResponse,
-  setPredictResponse,
-  predictAnswerLocked,
-  className,
-  offerBrowserTts,
-  canShowNextButton,
-  hasNextLevel,
-  useSecondaryFinishButton,
-  onContinueOrFinish,
-  bottomComponent,
-  noAnimation,
-}) => {
-  const vertical = layout === 'vertical';
-
-  const showSecondaryFinishButton = useSecondaryFinishButton && !hasNextLevel;
-
-  const useMessage =
-    showSecondaryFinishButton &&
-    queryParams('show-secondary-finish-button-question') === 'true'
-      ? commonI18n.finishMessage()
-      : message;
-
-  // The secondary finish button avoids a reappearance animation by not using
-  // the unique index.
-  const useMessageIndex = useSecondaryFinishButton ? undefined : messageIndex;
-
-  const feedbackRef = useRef<HTMLDivElement>(null);
-  const runButton = document.querySelector('#run-button');
-
-  useEffect(() => {
-    const checkRunButton = () => {
-      if (feedbackRef.current && runButton?.textContent === 'Run') {
-        feedbackRef.current.focus();
-        return true;
-      }
-      return false;
-    };
-
-    const observer = new MutationObserver(mutations => {
-      if (checkRunButton()) {
-        // Stop observing once the run button is found and has the text 'Run'
-        observer.disconnect();
-      }
-    });
-
-    observer.observe(document.body, {
-      childList: true,
-      subtree: true,
-    });
-
-    // Cleanup observer on component unmount
-    return () => {
-      observer.disconnect();
-    };
-  }, [useMessage, canShowNextButton, runButton?.textContent]);
-
-  return (
-    <div
-      id="instructions"
-      data-theme="Light"
-      className={classNames(
-        moduleStyles['instructions-' + theme],
-        vertical && moduleStyles.vertical,
-        'instructions',
-        className
-      )}
-    >
-      <div
-        id="instructions-panel"
-        className={classNames(
-          moduleStyles.item,
-          vertical && moduleStyles.itemVertical
-        )}
-      >
-        {text && (
-          <div
-            key={text}
-            id="instructions-text"
-            className={classNames(
-              moduleStyles['text-' + theme],
-              noAnimation && moduleStyles.noAnimation
-            )}
-          >
-            {offerBrowserTts && (
-              <TextToSpeech text={text} higherPosition={!!bottomComponent} />
-            )}
-            <div
-              id="instructions-text-content"
-              className={moduleStyles.textContent}
-            >
-              <div
-                className={
-                  offerBrowserTts
-                    ? moduleStyles.scrollingContentWithTTS
-                    : moduleStyles.scrollingContentWithoutTTS
-                }
-              >
-                <EnhancedSafeMarkdown
-                  markdown={text}
-                  className={moduleStyles.markdownText}
-                  handleInstructionsTextClick={handleInstructionsTextClick}
-                />
-                <PredictQuestion
-                  predictSettings={predictSettings}
-                  predictResponse={predictResponse}
-                  setPredictResponse={setPredictResponse}
-                  predictAnswerLocked={predictAnswerLocked}
-                />
-                {predictSettings?.isPredictLevel && (
-                  <InstructorsOnly>
-                    <div
-                      className={classNames(
-                        moduleStyles['message-' + theme],
-                        moduleStyles.predictSummary
-                      )}
-                    >
-                      <PredictSummary />
-                    </div>
-                  </InstructorsOnly>
-                )}
-              </div>
-              {bottomComponent && (
-                <div className={moduleStyles.bottomComponent}>
-                  {bottomComponent}
-                </div>
-              )}
-            </div>
-          </div>
-        )}
-        {(useMessage || canShowNextButton) && (
-          <div
-            key={useMessageIndex + ' - ' + useMessage}
-            id="instructions-feedback"
-            className={classNames(
-              moduleStyles.feedback,
-              showSecondaryFinishButton && moduleStyles.feedbackBottom
-            )}
-          >
-            <div
-              id="instructions-feedback-message"
-              className={classNames(
-                moduleStyles['message-' + theme],
-                noAnimation && moduleStyles.noAnimation
-              )}
-            >
-              {offerBrowserTts && useMessage && !canShowNextButton && (
-                <TextToSpeech text={useMessage} />
-              )}
-              {useMessage && (
-                <div ref={feedbackRef} tabIndex={-1}>
-                  <EnhancedSafeMarkdown
-                    markdown={useMessage}
-                    className={moduleStyles.markdownText}
-                    handleInstructionsTextClick={handleInstructionsTextClick}
-                  />
-                </div>
-              )}
-              {canShowNextButton && (
-                <Button
-                  id="instructions-continue-button"
-                  text={
-                    hasNextLevel ? commonI18n.continue() : commonI18n.finish()
-                  }
-                  onClick={onContinueOrFinish}
-                  className={moduleStyles.buttonInstruction}
-                  type={showSecondaryFinishButton ? 'secondary' : 'primary'}
-                  color={showSecondaryFinishButton ? 'black' : 'purple'}
-                />
-              )}
-            </div>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-};
-
 export default Instructions;
