+++ apps/src/lab2/views/components/InstructionsPanel.tsx
@@ -1,0 +1,235 @@
+import Button from '@code-dot-org/component-library/button';
+import classNames from 'classnames';
+import React, {useEffect, useRef} from 'react';
+
+import InstructorsOnly from '@cdo/apps/code-studio/components/InstructorsOnly';
+import {queryParams} from '@cdo/apps/code-studio/utils';
+import {LevelPredictSettings} from '@cdo/apps/lab2/levelEditors/types';
+import PredictQuestion from '@cdo/apps/lab2/views/components/PredictQuestion';
+import PredictSummary from '@cdo/apps/lab2/views/components/PredictSummary';
+import TextToSpeech from '@cdo/apps/lab2/views/components/TextToSpeech';
+import EnhancedSafeMarkdown from '@cdo/apps/templates/EnhancedSafeMarkdown';
+import {commonI18n} from '@cdo/apps/types/locale';
+
+import moduleStyles from '@cdo/apps/lab2/views/components/instructions.module.scss';
+
+interface InstructionsPanelProps {
+  /** Primary instructions text to display. */
+  text: string;
+  /** Optional message to display under the main text. This is typically a validation message. */
+  message?: string;
+  /** Key for rendering the optional message. A unique value ensures the appearance animation shows. */
+  messageIndex?: number;
+  /** If the instructions panel should be rendered vertically or horizontally. Defaults to vertical. */
+  layout?: 'vertical' | 'horizontal';
+  /** Display theme. Defaults to Dark. */
+  theme?: 'Dark' | 'Light';
+  /**
+   * A callback when the user clicks on clickable text.
+   */
+  handleInstructionsTextClick?: (id: string) => void;
+  predictSettings?: LevelPredictSettings;
+  predictResponse?: string;
+  setPredictResponse: (response: string) => void;
+  predictAnswerLocked: boolean;
+  /** Optional classname for the container */
+  className?: string;
+  offerBrowserTts?: boolean;
+  canShowNextButton: boolean;
+  hasNextLevel: boolean;
+  useSecondaryFinishButton: boolean;
+  onContinueOrFinish: () => void;
+  bottomComponent?: React.ReactNode;
+  noAnimation?: boolean;
+}
+
+/**
+ * Renders the instructions panel view, which is used within labs as well as to render a preview
+ * of the instructions panel on the level edit page.
+ */
+const InstructionsPanel: React.FunctionComponent<InstructionsPanelProps> = ({
+  text,
+  message,
+  messageIndex,
+  layout = 'vertical',
+  theme = 'Dark',
+  handleInstructionsTextClick,
+  predictSettings,
+  predictResponse,
+  setPredictResponse,
+  predictAnswerLocked,
+  className,
+  offerBrowserTts,
+  canShowNextButton,
+  hasNextLevel,
+  useSecondaryFinishButton,
+  onContinueOrFinish,
+  bottomComponent,
+  noAnimation,
+}) => {
+  const vertical = layout === 'vertical';
+
+  const showSecondaryFinishButton = useSecondaryFinishButton && !hasNextLevel;
+
+  const useMessage =
+    showSecondaryFinishButton &&
+    queryParams('show-secondary-finish-button-question') === 'true'
+      ? commonI18n.finishMessage()
+      : message;
+
+  // The secondary finish button avoids a reappearance animation by not using
+  // the unique index.
+  const useMessageIndex = useSecondaryFinishButton ? undefined : messageIndex;
+
+  const feedbackRef = useRef<HTMLDivElement>(null);
+  const runButton = document.querySelector('#run-button');
+
+  useEffect(() => {
+    const checkRunButton = () => {
+      if (feedbackRef.current && runButton?.textContent === 'Run') {
+        feedbackRef.current.focus();
+        return true;
+      }
+      return false;
+    };
+
+    const observer = new MutationObserver(mutations => {
+      if (checkRunButton()) {
+        // Stop observing once the run button is found and has the text 'Run'
+        observer.disconnect();
+      }
+    });
+
+    observer.observe(document.body, {
+      childList: true,
+      subtree: true,
+    });
+
+    // Cleanup observer on component unmount
+    return () => {
+      observer.disconnect();
+    };
+  }, [useMessage, canShowNextButton, runButton?.textContent]);
+
+  return (
+    <div
+      id="instructions"
+      data-theme="Light"
+      className={classNames(
+        moduleStyles['instructions-' + theme],
+        vertical && moduleStyles.vertical,
+        'instructions',
+        className
+      )}
+    >
+      <div
+        id="instructions-panel"
+        className={classNames(
+          moduleStyles.item,
+          vertical && moduleStyles.itemVertical
+        )}
+      >
+        {text && (
+          <div
+            key={text}
+            id="instructions-text"
+            className={classNames(
+              moduleStyles['text-' + theme],
+              noAnimation && moduleStyles.noAnimation
+            )}
+          >
+            {offerBrowserTts && (
+              <TextToSpeech text={text} higherPosition={!!bottomComponent} />
+            )}
+            <div
+              id="instructions-text-content"
+              className={moduleStyles.textContent}
+            >
+              <div
+                className={
+                  offerBrowserTts
+                    ? moduleStyles.scrollingContentWithTTS
+                    : moduleStyles.scrollingContentWithoutTTS
+                }
+              >
+                <EnhancedSafeMarkdown
+                  markdown={text}
+                  className={moduleStyles.markdownText}
+                  handleInstructionsTextClick={handleInstructionsTextClick}
+                />
+                <PredictQuestion
+                  predictSettings={predictSettings}
+                  predictResponse={predictResponse}
+                  setPredictResponse={setPredictResponse}
+                  predictAnswerLocked={predictAnswerLocked}
+                />
+                {predictSettings?.isPredictLevel && (
+                  <InstructorsOnly>
+                    <div
+                      className={classNames(
+                        moduleStyles['message-' + theme],
+                        moduleStyles.predictSummary
+                      )}
+                    >
+                      <PredictSummary />
+                    </div>
+                  </InstructorsOnly>
+                )}
+              </div>
+              {bottomComponent && (
+                <div className={moduleStyles.bottomComponent}>
+                  {bottomComponent}
+                </div>
+              )}
+            </div>
+          </div>
+        )}
+        {(useMessage || canShowNextButton) && (
+          <div
+            key={useMessageIndex + ' - ' + useMessage}
+            id="instructions-feedback"
+            className={classNames(
+              moduleStyles.feedback,
+              showSecondaryFinishButton && moduleStyles.feedbackBottom
+            )}
+          >
+            <div
+              id="instructions-feedback-message"
+              className={classNames(
+                moduleStyles['message-' + theme],
+                noAnimation && moduleStyles.noAnimation
+              )}
+            >
+              {offerBrowserTts && useMessage && !canShowNextButton && (
+                <TextToSpeech text={useMessage} />
+              )}
+              {useMessage && (
+                <div ref={feedbackRef} tabIndex={-1}>
+                  <EnhancedSafeMarkdown
+                    markdown={useMessage}
+                    className={moduleStyles.markdownText}
+                    handleInstructionsTextClick={handleInstructionsTextClick}
+                  />
+                </div>
+              )}
+              {canShowNextButton && (
+                <Button
+                  id="instructions-continue-button"
+                  text={
+                    hasNextLevel ? commonI18n.continue() : commonI18n.finish()
+                  }
+                  onClick={onContinueOrFinish}
+                  className={moduleStyles.buttonInstruction}
+                  type={showSecondaryFinishButton ? 'secondary' : 'primary'}
+                  color={showSecondaryFinishButton ? 'black' : 'purple'}
+                />
+              )}
+            </div>
+          </div>
+        )}
+      </div>
+    </div>
+  );
+};
+
+export default InstructionsPanel;
